+++ R/raster-info.R
@@ -40,6 +40,8 @@ sds_boilerplate_checks <- function(x, sds = NULL) {
 #' \item{projection}{text version of map projection parameter string}
 #' \item{bands}{number of bands in the dataset}
 #' \item{proj4}{not implemented}
+#' \item{nodata_value}{not implemented}
+#' \item{overviews}{the number and size of any available overviews}
 #' }
 #'
 #' On access vapour functions will report on the existence of subdatasets while
@@ -104,6 +106,12 @@ sds_boilerplate_checks <- function(x, sds = NULL) {
 #' Calculation of 'minmax' can take a significant amount of time, so it's not done by default. Use
 #' 'minmax = TRUE' to do it. (It does perform well, but may be prohibitive for very large or remote sources.)
 #' @seealso vapour_sds_info
+#'
+#' @section Overviews:
+#'
+#' If there are no overviews this element will simply be a single-element vector
+#' of value 0. If there are overviews, the first value will give the number of overviews and
+#' their dimensions will be listed as pairs of x,y values.
 #'
 #' @param x data source string (i.e. file name or URL or database connection string)
 #' @param sds a subdataset number, if necessary