+++ tests/testthat/test-raster-info.R
@@ -35,8 +35,12 @@ test_that("raster info works", {
   expect_error(vapour_sds_names(), 'argument "x" is missing, with no default')
   expect_error(vapour_sds_names(""), 'cannot open dataset')
 
+  f1 <- system.file("extdata/volcano_overview.tif", package = "vapour", mustWork = TRUE)
+  f2 <- system.file("extdata/volcano.tif", package = "vapour", mustWork = TRUE)
 
-
+  expect_true(vapour_raster_info(f2)$overviews == 0L)
+  expect_equal(vapour_raster_info(f1)$overviews, ## Overviews: 31x44, 16x22, 8x11, 4x6
+               c(4L, 31L, 44L, 16L, 22L, 8L, 11L, 4L, 6L))
   ## these are different on Windows and Linux ... because linux sees more SDS
   ## expect_error(sds_boilerplate_checks(f, 0), "sds must be 1 at minimum")
     # expect_message(sds_boilerplate_checks(f), "subdataset \\(variable\\) used is '//f16'")